requirements.txt
# Requirements file optimized for macOS (especially Apple Silicon)
# Core FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.7  # PostgreSQL adapter for Python
alembic>=1.12.1  # Database migrations

# Configuration and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# HTTP client for external APIs
httpx>=0.25.2

# File processing
PyPDF2>=3.0.1
pdfminer.six>=20231228
docx2txt>=0.8
python-magic>=0.4.27

# NLP and text processing (with macOS-compatible alternatives)
# Using newer spacy with pre-built wheels that work on macOS
spacy>=3.7.0
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5

# Fuzzy string matching (alternative to rapidfuzz that works better on macOS)
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1

# Data processing and utilities
pandas>=2.1.4
numpy>=1.24.0,<2.0.0  # Pin to avoid compatibility issues
scikit-learn>=1.3.2
joblib>=1.3.2

# File type detection
python-magic>=0.4.27

# Logging and monitoring
structlog>=23.2.0

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Email functionality
aiosmtplib>=3.0.1
jinja2>=3.1.2

# Environment and configuration
python-decouple>=3.8

# Async utilities (commented out problematic packages)
# asyncio-mqtt>=0.16.1
aiofiles>=23.2.1

# Additional utilities for the project
click>=8.1.7
rich>=13.7.0
typer>=0.9.0
